Convert size_t, count_t and index_t to 64-bits on 64-bit architectures. Change some pfn_t usages to count_t, size_t and index_t.

+++ generic/include/typedefs.h
@@ -34,9 +34,9 @@
 
 typedef short bool;
 
-typedef unsigned int size_t;
-typedef unsigned int count_t;
-typedef unsigned int index_t;
+typedef unsigned long size_t;
+typedef unsigned long count_t;
+typedef unsigned long index_t;
 
 typedef struct config config_t;
 typedef struct cpu_info cpu_info_t;
